% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_centre.R
\name{coerce_centre}
\alias{coerce_centre}
\title{Convert an object to a point in a suitable format}
\usage{
coerce_centre(centre, ...)
}
\arguments{
\item{centre}{Object defining the centre point.}

\item{...}{Unused in the default implementation.}
}
\value{
A \code{sfc} object containins a single \code{\link[sf:st]{sf::st_point}} feature
and with an appropriate CRS set.
}
\description{
This is an internal helper used by the shape definition functions
to process their \code{centre} parameters, where present.
It is not intended to be used directly.
Instead, it allows extensions to integrate with the shape
functions so that they can accept new definitions.
}
