% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_sf_inset.R
\name{coord_sf_inset}
\alias{coord_sf_inset}
\title{Specify an inset configuration for the whole plot}
\usage{
coord_sf_inset(inset, ...)
}
\arguments{
\item{inset}{Inset configuration; see \code{\link[=configure_inset]{configure_inset()}}.}

\item{...}{Arguments passed to \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}}}
}
\description{
This allows a default inset configuration to be provided to avoid having to
repeat it for each layer. Any layer that is inset-aware can use this as the
default configuration if none is specifically provided to that layer.
}
