% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_inset_frame.R
\name{geom_inset_frame}
\alias{geom_inset_frame}
\title{Add a frame and burst lines for an inset.}
\usage{
geom_inset_frame(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "sf",
  position = "identity",
  ...,
  inset = NULL,
  na.rm = FALSE,
  source.params = list(),
  target.params = list(),
  lines.params = list(),
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}}

\item{inset}{Inset configuration; see \code{\link[=configure_inset]{configure_inset()}}.}

\item{source.params, target.params, lines.params}{Override the aesthetics of the
inset source, target, and lines respectively. The value should be a list
named by the aesthetics, and the values should be scalars of length one.}
}
\value{
A ggplot layer holding the inset frame.
}
\description{
Add a frame and burst lines for an inset.
}
\section{Limitation}{

Currently the inset configuration must be passed specifically to this layer
and the frame will be built at call time instead of when the rest of the plot
is built.
}

\examples{
library(sf)
library(ggplot2)

nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
cfg <- configure_inset(
  centre = st_sfc(st_point(c(-82, 35)), crs = 4326),
  scale = 2,
  translation = c(0, -300),
  radius = 50,
  units = "mi")

ggplot() + geom_inset_frame(inset = cfg, data = nc)
}
