% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_inset.R
\docType{data}
\name{geom_sf_inset}
\alias{geom_sf_inset}
\alias{GeomSfInset}
\title{Visualise sf objects with insets}
\usage{
geom_sf_inset(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "sf",
  position = "identity",
  ...,
  inset = NULL,
  map_base = "normal",
  map_inset = "auto",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}}

\item{inset}{Inset configuration; see \code{\link[=configure_inset]{configure_inset()}}.}

\item{map_base}{Controls the layer with the base map. Possible values are
\code{"normal"} to create a layer as though the inset were not specified,
\code{"clip"} to create a layer with the inset viewport cut out, and
\code{"none"} to prevent the insertion of a layer for the base map.}

\item{map_inset}{Controls the layer with the inset map. Possible values are
\code{"auto"} to choose the behaviour based on whether \code{inset} is specified,
\code{"normal"} to create a layer with the viewport cut out and transformed, and
\code{"none"} to prevent the insertion og a layer for the viewport map.}
}
\value{
A ggplot layer similar to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} but transformed according to the
inset configuration.
}
\description{
This is a wrapper around \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} that assists with creating map
insets.
First, configure an inset using \code{\link[=configure_inset]{configure_inset()}}, then pass around the
configuration object using the \code{inset} parameter.
After specifying all your usual geoms, use \code{\link[=geom_inset_frame]{geom_inset_frame()}} to add a frame
around the inset that connects it to the main map.
}
\details{
Internally this works by creating two layers: one for the base map, and one
for the inset. These can be separately controlled by the \code{map_base} and
\code{map_inset} parameters. If \code{inset} is not specified, this geom will instead
behave like \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.

When an inset is configured, the default creates both base and inset layers
using the same aesthetic mapping and params:

\if{html}{\out{<div class="sourceCode">}}\preformatted{cfg <- configure_inset(...)
geom_sf_inset(aes(...), inset = cfg, ...)
}\if{html}{\out{</div>}}

You can alternatively specify the two layers separately:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# draw the base map only (both versions are equivalent):
geom_sf(aes(...), ...)
geom_sf_inset(aes(...), inset = cfg, map_inset = "none", ...)
# draw the inset map only:
geom_sf_inset(aes(...), inset = cfg, map_base = "none", ...)
}\if{html}{\out{</div>}}
}
\examples{
library(sf)
library(ggplot2)

nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
cfg <- configure_inset(
  centre = st_sfc(st_point(c(-80, 35.5)), crs = 4326),
  scale = 1.5,
  translation = c(-50, -140),
  radius = 50,
  units = "mi")

ggplot(nc) +
  geom_sf_inset(aes(fill = AREA), inset = cfg) +
  geom_inset_frame(inset = cfg) +
  coord_sf()
}
\keyword{datasets}
